Squashed commit of the following:
commit 8c1007e29723d8a91d81a808958b9dcd66c36aa9
Author: Brandon Fong <55638760+BrandonMFong@users.noreply.github.com>
Date:   Sun May 24 16:23:27 2020 -0700

    incorporated into time.ps1

commit 2cab5d0b1418ec62b2bca44b43c993abe7340b5c
Author: Brandon Fong <55638760+BrandonMFong@users.noreply.github.com>
Date:   Sun May 24 16:07:14 2020 -0700

    Time duration finished

    TODO incorporate into time.ps1

commit 58983d841d9143e0b1fd05dffc5a07fd4a1030ec
Author: Brandon Fong <55638760+BrandonMFong@users.noreply.github.com>
Date:   Sun May 24 15:40:01 2020 -0700

    [Code] Editting query to insert time in/out data

commit 569248fca48fa33c654dbd8bb9fdea8a046d8f13
Author: Brandon Fong <55638760+BrandonMFong@users.noreply.github.com>
Date:   Sun May 24 15:19:35 2020 -0700

    Method for Time in/out finished

+++ Functions/TIME.ps1
@@ -22,35 +22,48 @@ Param
 	[Alias('C')][Switch]$Clear, 
 	[Alias('A')][Switch]$Archive
 )
-Push-Location C:\Users\brandon.fong\Brandon.Fong\DOCUMENTS\User;
-	if(!(Test-Path .\TimeStamp.csv)){New-Item TimeStamp.csv}
-	if($Login)
+[Calendar]$var = GetObjectByClass('Calendar');
+
+if([string]::IsNullOrEmpty($var.TimeStampFilePath))# Database config
+{
+	if($Login){$var.TimeIn();}
+	if($Logout){$var.TimeOut();}
+	if($View){$var.GetTimeStampDuration();}
+}
+
+# todo finish
+else 
+{
+	Push-Location C:\Users\brandon.fong\Brandon.Fong\DOCUMENTS\User;
+		if(!(Test-Path .\TimeStamp.csv)){New-Item TimeStamp.csv}
+		if($Login)
+			{
+				$i = Get-Date 
+				Add-Content C:\Users\brandon.fong\Brandon.Fong\DOCUMENTS\User\TimeStamp.csv "Login, $i"
+			}
+		if($Logout)
+			{
+				$i = Get-Date 
+				Add-Content C:\Users\brandon.fong\Brandon.Fong\DOCUMENTS\User\TimeStamp.csv "LogOut, $i"
+			}
+		if($View)
+			{
+				Get-Content C:\Users\brandon.fong\Brandon.Fong\DOCUMENTS\User\TimeStamp.csv
+			}
+		if($Clear) # Only call this from within the script
+			{
+				Clear-Content C:\Users\brandon.fong\Brandon.Fong\DOCUMENTS\User\TimeStamp.csv
+			}
+		if($Archive)
 		{
-			$i = Get-Date 
-			Add-Content C:\Users\brandon.fong\Brandon.Fong\DOCUMENTS\User\TimeStamp.csv "Login, $i"
+			if(!(Test-Path .\log_archive\)){mkdir .\log_archive\;}
+			$date = Get-Date;
+			$file = "Archive_" + $date.Month.ToString() + $date.Day.ToString() + $date.Year.ToString() + ".csv";
+			New-Item $file;
+			$Temp = Get-Content C:\Users\brandon.fong\Brandon.Fong\DOCUMENTS\User\TimeStamp.csv;
+			Add-Content $file $Temp;
+			Move-Item $file .\log_archive;
+			Time -Clear; #calling itself to clear its contents
 		}
-	if($Logout)
-		{
-			$i = Get-Date 
-			Add-Content C:\Users\brandon.fong\Brandon.Fong\DOCUMENTS\User\TimeStamp.csv "LogOut, $i"
-		}
-	if($View)
-		{
-			Get-Content C:\Users\brandon.fong\Brandon.Fong\DOCUMENTS\User\TimeStamp.csv
-		}
-	if($Clear) # Only call this from within the script
-		{
-			Clear-Content C:\Users\brandon.fong\Brandon.Fong\DOCUMENTS\User\TimeStamp.csv
-		}
-	if($Archive)
-	{
-		if(!(Test-Path .\log_archive\)){mkdir .\log_archive\;}
-		$date = Get-Date;
-		$file = "Archive_" + $date.Month.ToString() + $date.Day.ToString() + $date.Year.ToString() + ".csv";
-		New-Item $file;
-		$Temp = Get-Content C:\Users\brandon.fong\Brandon.Fong\DOCUMENTS\User\TimeStamp.csv;
-		Add-Content $file $Temp;
-		Move-Item $file .\log_archive;
-		Time -Clear; #calling itself to clear its contents
-	}
-Pop-Location
+	Pop-Location
+}
